.history/requirements_20251111150633.txt
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.29.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.6.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
pypdf==3.17.4
